requirements.txt

Flask
pdfplumber
pytesseract
python-dotenv
Werkzeug
Pillow
